requirements.txt
streamlit==1.33.0
pymupdf==1.23.26
sentence-transformers==2.6.1
faiss-cpu==1.7.4
groq==0.9.0
numpy==1.26.4
torch==2.1.2
